requirements.txt
aniso8601==7.0.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
black==23.10.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
Django==4.2.6
django-cors-headers==4.3.0
django-redis==5.4.0
django-vite==2.1.3
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
hiredis==2.2.3
idna==3.4
inertia-django==0.5.3
more-itertools==9.1.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
promise==2.3
pycparser==2.21
pydantic==2.0.3
pydantic-core==2.3.0
python-ulid==1.1.0
redis==4.6.0
redis-om==0.2.1
requests==2.31.0
Rx==1.6.3
singledispatch==4.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
types-pyOpenSSL==23.3.0.0
types-redis==4.6.0.8
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
whitenoise==6.6.0
